+++ app/src/main/java/com/example/recipekeeper/ui/sharedcomposable/SectionTitle.kt
@@ -1,0 +1,13 @@
+package com.example.recipekeeper.ui.sharedcomposable
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SectionTitle(title: String) {
+    Text(
+        text = title,
+        style = MaterialTheme.typography.titleMedium
+    )
+}
